fix(sessions): stabilize SQLite archive CI
