+++ src/config/sessions/session-accessor.sqlite-delete-snapshot.types.ts
@@ -1,0 +1,9 @@
+export type SessionStateDeleteSnapshot = {
+  acpParentStreamEventCount: number;
+  generation: string | null;
+  lastSeq: number | null;
+  sessionKey: string | null;
+  sessionUpdatedAt: number | null;
+  trajectoryLastSeq: number | null;
+  transcriptUpdatedAt: number | null;
+};
